# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CreateNotificationRequest:
    properties:
      link:
        type: string
      message:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: integer
    required:
    - title
    - type
    - user_id
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 400
        example: 400
        type: integer
      details:
        description: Field-level validation error details (for validation errors)
        items:
          $ref: '#/definitions/models.FieldError'
        type: array
      error:
        description: |-
          Error type/category
          example: Bad Request
        example: Bad Request
        type: string
      message:
        description: |-
          Detailed error message
          example: Invalid email format
        example: Invalid email format
        type: string
      request_id:
        description: |-
          Request ID for tracing and debugging
          example: 550e8400-e29b-41d4-a716-446655440000
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  models.FieldError:
    properties:
      code:
        description: |-
          Error code for programmatic handling
          example: email
        type: string
      field:
        description: |-
          The field name that failed validation
          example: email
        type: string
      message:
        description: |-
          Human-readable error message
          example: Invalid email format
        type: string
      value:
        description: The rejected value (excluded for sensitive fields)
    type: object
...
  models.NotificationResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      link:
        type: string
      message:
        type: string
      read:
        type: boolean
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  models.NotificationsResponse:
    properties:
      notifications:
        items:
          $ref: '#/definitions/models.NotificationResponse'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      unread:
        type: integer
    type: object
...
  models.SuccessResponse:
    properties:
      data:
        description: Response data (varies by endpoint)
      message:
        description: |-
          Success message
          example: User created successfully
        example: User created successfully
        type: string
      success:
        description: |-
          Operation success status
          example: true
        example: true
        type: boolean
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/your-org/react-golang-starter-kit/issues
  description: A comprehensive REST API for the React Go Starter Kit application built
    with Fiber, GORM, and PostgreSQL. This API provides secure user authentication,
    user management, and system health monitoring.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/your-org/react-golang-starter-kit
  title: React Go Starter Kit API
  version: 1.0.0
paths:
...
  /api/admin/notifications:
    post:
      consumes:
      - application/json
      description: Creates a notification for a specific user
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create notification (admin)
      tags:
      - Notifications
...
  /api/notifications:
    get:
      description: Returns paginated list of notifications for the authenticated user
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: per_page
        type: integer
      - description: Filter to only unread notifications
        in: query
        name: unread
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NotificationsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get user notifications
      tags:
      - Notifications
  /api/notifications/{id}:
    delete:
      description: Deletes a specific notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete notification
      tags:
      - Notifications
  /api/notifications/{id}/read:
    post:
      description: Marks a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Mark notification as read
      tags:
      - Notifications
  /api/notifications/read-all:
    post:
      description: Marks all unread notifications as read for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Mark all notifications as read
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Format: `Authorization:
      Bearer {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"